requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.10.1
Pillow==10.1.0
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
Flask==2.3.3
typing-extensions==4.8.0
face-recognition==1.3.0
pymongo==4.5.0
fastapi==0.95.2
python-multipart==0.0.6
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
motor==3.3.1
ultralytics==8.1.2 